// src/app/generated/cite.api/model/groupMembership.ts
/*
 Copyright 2026 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Cite API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GroupMembership { 
    id?: string;
    /**
     * ID of the group.
     */
    groupId?: string;
    /**
     * Id of the User.
     */
    userId?: string;
}